Display disabled EC card when Google ads is disconnected.
